+++ poc_ann/FannManagerNormal.h
@@ -14,10 +14,10 @@ public:
 	void train();
 	FannManagerNormal();
 	~FannManagerNormal();
-	friend int logOut(FANN::neural_net&, FANN::training_data&, unsigned int, unsigned int, float, unsigned int, void*);
+	friend int logOut(FANN::neural_net&, FANN::training_data&, unsigned int, unsigned int, double, unsigned int, void*);
 };
 
-int logOut(FANN::neural_net &net, FANN::training_data &train, unsigned int max_epochs, unsigned int epochs_between_reports, float desired_error, unsigned int epochs, void *user_data);
+int logOut(FANN::neural_net &net, FANN::training_data &train, unsigned int max_epochs, unsigned int epochs_between_reports, double desired_error, unsigned int epochs, void *user_data);
 
 
 struct EscapeFM
